+++ R/acled.api.R
@@ -7,6 +7,8 @@
 #' the research work flow. \cr \cr
 #' When using this package, you acknowledge that you have read ACLED's terms and
 #' conditions of use, and that you agree with their attribution requirements.
+#' @param email.address character string. Supply the email address that you registered with [ACLED access](https://developer.acleddata.com/).
+#' @param access.key character string. Supply your ACLED access key.
 #' @param country character vector. Supply one or more country names to narrow down which events should be retrieved. See the details
 #' below for information on how the arguments "country" and "region" interact.
 #' @param region numeric or character vector. Supply one or more region codes (numeric) or region names (character)
@@ -45,12 +47,16 @@
 #' Clionadh Raleigh, Andrew Linke, Havard Hegre and Joakim Karlsen. 2010.
 #' "Introducing ACLED-Armed Conflict Location and Event Data." _Journal of Peace Research_ 47 (5): 651-660.
 #' @examples
-#' my.data.frame1 <- acled.api(region = c(1,7),
+#' my.data.frame1 <- acled.api(email.address = "your.email.address",
+#' access.key = "your.access.key",
+#' region = c(1,7),
 #' start.date = "2018-11-01",
 #' end.date = "2018-11-31")
 #' head(my.data.frame1)
 #'
-#' my.data.frame2 <- acled.api(region = c(1,7),
+#' my.data.frame2 <- acled.api(email.address = "your.email.address",
+#' access.key = "your.access.key",
+#' region = c(1,7),
 #' start.date = "2018-11-01",
 #' end.date = "2018-11-31",
 #' add.variables = c("geo_precision", "time_precision"))
@@ -58,6 +64,8 @@
 #' @export
 #'
 acled.api <- function(
+  email.address = Sys.getenv("EMAIL_ADDRESS"),
+  access.key = Sys.getenv("ACCESS_KEY"),
   country = NULL,
   region = NULL,
   start.date = NULL,
@@ -67,8 +75,23 @@ acled.api <- function(
   dyadic = FALSE,
   other.query = NULL){
 
-  # accept terms
-  terms <- "read?terms=accept&limit=0"
+
+  # access key
+  if ( (is.null(access.key) | !is.character(access.key) | access.key=="") == TRUE ) {
+    stop('ACLED requires an access key, which needs to be supplied to this argument as a character string.
+    You can request an access key by registering on https://developer.acleddata.com/.', call. = FALSE)
+  }
+  if( (is.character(access.key) & access.key!="") == TRUE){
+  access.key1 <- paste0("key=", access.key)}
+
+  # email address
+  if ( (is.null(email.address) | !is.character(email.address) | email.address=="") == TRUE ) {
+    stop('ACLED requires an email address for access, which needs to be supplied to this argument as a character string.
+    Use the email address you provided when registering on https://developer.acleddata.com/.', call. = FALSE)
+  }
+  if( (is.character(email.address) & email.address!="") == TRUE){
+  email.address1 <- paste0("&email=", email.address)}
+
 
   # country argument
   if ( (!is.null(country) & !is.character(country)) == TRUE ) {
@@ -156,7 +179,7 @@ acled.api <- function(
 
   # ACLED ping
   acled.url.ping <- tryCatch(
-    httr::GET("https://api.acleddata.com/acled/read?terms=accept&limit=1&?export_type=monadic"),
+    httr::GET("https://api.acleddata.com/"),
     error = function(e) e )
   if( any(class(acled.url.ping) == "error") ) {
     message("The resource api.acleddata.com cannot be reached. \n
@@ -168,7 +191,8 @@ acled.api <- function(
 
 
   # GET call
-  url <- paste0("https://api.acleddata.com/acled/", terms, dyadic1, time.frame1, variables, country1, region1, other.query1)
+  url <- paste0("https://api.acleddata.com/acled/read/?",
+                access.key1, email.address1, "&limit=0", dyadic1, time.frame1, variables, country1, region1, other.query1)
 
   response <- httr::GET(url)
   if ( exists("response")==FALSE ) {
@@ -187,7 +211,7 @@ acled.api <- function(
   json.content <- jsonlite::fromJSON( httr::content(response, "text", encoding = 'UTF-8'),
                             simplifyVector = FALSE)
   if(!json.content$success){
-    message(paste0("GET request wasn't successful: ", json.content$error$message))
+    message(paste0("GET request wasn't successful. The API returned status ", json.content$error$status, ": ", json.content$error$message, "."))
     return(NULL)
   }
   json.content <- json.content$data